refactor: extract all firebase and nx-translate dependencies into one separate module
